Move relevant functions back to config.py
config.py now holds AMQP-related code and extension registration
code as well, because these steps should only be relevant during
cse installation.
