+++ container_service_extension/utils.py
@@ -12,14 +12,10 @@ import ssl
 import stat
 import string
 import sys
-import traceback
 from urllib.parse import urlparse
 
 import click
-import pika
 from cachetools import LRUCache
-from pyvcloud.vcd.amqp import AmqpService
-from pyvcloud.vcd.api_extension import APIExtension
 from pyvcloud.vcd.client import BasicLoginCredentials
 from pyvcloud.vcd.client import Client
 from pyvcloud.vcd.exceptions import EntityNotFoundException
@@ -254,108 +250,6 @@ def create_and_share_catalog(org, catalog_name, catalog_desc=''):
     return org.get_catalog(catalog_name)
 
 
-def register_extension(client, ext_name, exchange_name, patterns=None):
-    """Registers an API extension in vCD.
-
-    :param pyvcloud.vcd.client.Client client:
-    :param str ext_name: extension name.
-    :param str exchange_name: AMQP exchange name.
-    :param list patterns: list of urls that map to this API extension.
-    """
-    ext = APIExtension(client)
-    if patterns is None:
-        patterns = [
-            f'/api/{ext_name}',
-            f'/api/{ext_name}/.*',
-            f'/api/{ext_name}/.*/.*'
-        ]
-
-    ext.add_extension(ext_name, ext_name, ext_name, exchange_name, patterns)
-    click.secho(f"Registered extension {ext_name} as an API extension in vCD",
-                fg='green')
-
-
-def configure_vcd_amqp(client, exchange_name, host, port, prefix,
-                       ssl_accept_all, use_ssl, vhost, username, password):
-    """Configures vCD AMQP settings/exchange using parameter values.
-
-    :param pyvcloud.vcd.client.Client client:
-    :param str exchange_name: name of exchange.
-    :param str host: AMQP host name.
-    :param str password: AMQP password.
-    :param int port: AMQP port.
-    :param str prefix:
-    :param bool ssl_accept_all:
-    :param bool use_ssl: Enable ssl.
-    :param str username: AMQP username.
-    :param str vhost: AMQP vhost.
-
-    :raises Exception: if could not set AMQP configuration.
-    """
-    amqp_service = AmqpService(client)
-    amqp = {
-        'AmqpExchange': exchange_name,
-        'AmqpHost': host,
-        'AmqpPort': port,
-        'AmqpPrefix': prefix,
-        'AmqpSslAcceptAll': ssl_accept_all,
-        'AmqpUseSSL': use_ssl,
-        'AmqpUsername': username,
-        'AmqpVHost': vhost
-    }
-
-    # This block sets the AMQP setting values on the
-    # vCD "System Administration Extensibility page"
-    result = amqp_service.test_config(amqp, password)
-    click.secho(f"AMQP test settings, result: {result['Valid'].text}",
-                fg='yellow')
-    if result['Valid'].text == 'true':
-        amqp_service.set_config(amqp, password)
-        click.secho("Updated vCD AMQP configuration", fg='green')
-    else:
-        msg = "Couldn't set vCD AMQP configuration"
-        click.secho(msg, fg='red')
-        # TODO replace raw exception with specific
-        raise Exception(msg)
-
-
-def create_amqp_exchange(exchange_name, host, port, vhost, use_ssl,
-                         username, password):
-    """Creates the specified AMQP exchange if it does not exist.
-
-    If specified AMQP exchange exists already, does nothing.
-
-    :param str exchange_name: The AMQP exchange name to check for or create.
-    :param str host: AMQP host name.
-    :param str password: AMQP password.
-    :param int port: AMQP port number.
-    :param bool use_ssl: Enable ssl.
-    :param str username: AMQP username.
-    :param str vhost: AMQP vhost.
-
-    :raises Exception: if AMQP exchange could not be created.
-    """
-    click.secho(f"Checking for AMQP exchange '{exchange_name}'", fg='yellow')
-    credentials = pika.PlainCredentials(username, password)
-    parameters = pika.ConnectionParameters(host, port, vhost, credentials,
-                                           ssl=use_ssl, connection_attempts=3,
-                                           retry_delay=2, socket_timeout=5)
-    try:
-        connection = pika.BlockingConnection(parameters)
-        click.secho(f"Connected to AMQP server: {host}:{port}", fg='green')
-        channel = connection.channel()
-        channel.exchange_declare(exchange=exchange_name,
-                                 exchange_type=EXCHANGE_TYPE,
-                                 durable=True, auto_delete=False)
-    except Exception:  # TODO replace with specific exception
-        LOGGER.error(traceback.format_exc())
-        click.secho(f"Couldn't create exchange '{exchange_name}'", fg='red')
-        raise
-    finally:
-        connection.close()
-    click.secho(f"AMQP exchange '{exchange_name}' is ready", fg='green')
-
-
 def bool_to_msg(b):
     if b:
         return 'success'
@@ -413,29 +307,3 @@ def get_vsphere(config, vapp, vm_name):
                 cache[vm_id]['password'], cache[vm_id]['port'])
 
     return v
-
-
-# unused functions, unsure what to do with these
-# import hashlib
-# import random
-# import socket
-# import ssl
-# import string
-# def hex_chunks(s):
-#     return [s[i:i + 2] for i in range(0, len(s), 2)]
-
-
-# def get_thumbprint(host, port):
-#     sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-#     sock.settimeout(10)
-#     wrappedSocket = ssl.wrap_socket(sock)
-#     wrappedSocket.connect((host, port))
-#     der_cert_bin = wrappedSocket.getpeercert(True)
-#     thumb_sha1 = hashlib.sha1(der_cert_bin).hexdigest()
-#     wrappedSocket.close()
-#     return ':'.join(map(str, hex_chunks(thumb_sha1))).upper()
-
-
-# def random_word(length):
-#     letters = string.ascii_lowercase
-#     return ''.join(random.choice(letters) for i in range(length))
